+++ Assets/Scripts/UI/MainMenuHandler.cs
@@ -12,8 +12,8 @@ public class MainMenuHandler : MonoBehaviour {
 
     void Start() {
         // scene index from build settings
-        this.play_button.onClick.AddListener(() => SceneManager.LoadScene(Scenes.LEVELSELECTION));
-        this.options_button.onClick.AddListener(() => SceneManager.LoadScene(Scenes.OPTIONS));
+        this.play_button.onClick.AddListener(() => SceneManager.LoadScene((int)Scenes.LEVELSELECTION));
+        this.options_button.onClick.AddListener(() => SceneManager.LoadScene((int)Scenes.OPTIONS));
         this.quit_button.onClick.AddListener(() => Application.Quit());
     }
 }
